+++ src/util.js
@@ -1,5 +1,6 @@
-const fs = require("fs");
+// eslint-disable-next-line import/no-unresolved
 const vscode = require("vscode");
+const fs = require("fs");
 
 /**
  * If the file exists, it does nothing. If  the file does not exist, it will create a new file with an empty object.
@@ -26,7 +27,18 @@ async function askForLanguage() {
   return result;
 }
 
+function escapeStringRegexp(string) {
+  if (typeof string !== "string") {
+    throw new TypeError("Expected a string");
+  }
+
+  // Escape characters with special meaning either inside or outside character sets.
+  // Use a simple backslash escape when it’s always valid, and a `\xnn` escape when the simpler form would be disallowed by Unicode patterns’ stricter grammar.
+  return string.replace(/[|\\{}()[\]^$+*?.]/g, "\\$&").replace(/-/g, "\\x2d");
+}
+
 module.exports = {
   createFileIfDoesNotExist,
   askForLanguage,
+  escapeStringRegexp,
 };