+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.14.0] - 2021-04-22
+
+### Added
+
+- Added *edit* action button to tables.
+
+### Changed
+
+- Updated README with new demo GIF and images.
+
 ## [0.13.0] - 2021-03-31
 
 ### Changed